phase-2/backend/requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlmodel>=0.0.14
asyncpg>=0.29.0
alembic>=1.13.0
greenlet>=3.0.0

# Environment
python-dotenv>=1.0.0

# Development/Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
